refactor(message): move common enums and types into separate file
