+++ src/common/enums.ts
@@ -1,0 +1,6 @@
+export enum MessageEnum {
+  NEW = 'new',
+  MEMBERS = 'members',
+  MESSAGE = 'message',
+  OUT = 'out',
+}
