+++ src/common/types.ts
@@ -1,3 +1,5 @@
+import { MessageEnum } from './enums';
+
 export interface Agenda {
   _id: string;
   title: string;
@@ -7,4 +9,11 @@ export interface Agenda {
   votesCountMap: Record<string, number>;
   userChoice: string | null;
   expires: string; // ISO Date String
+}
+
+export interface MessageType {
+  type: MessageEnum;
+  payload: string;
+  issuer?: string;
+  date?: string;
 }
